+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 For packages corresponding to these changes, see the [releases page](https://github.com/Justin-Credible/nintendo-vs-frontend/releases).
 
 ## [v1.2.0]
-TBD
+TBD - In Progress
+### Changed
+- Ability to change menu sound effects volume via `sfxVolume` in `config.yml`
 
 ## [v1.1.0](https://github.com/Justin-Credible/nintendo-vs-frontend/releases/tag/v1.1.0) - 2019-06-09
 ### Added